file writing works, dates are correct
I fixed the problem about only the last element showing up. It was because I was overwriting what I was previously writing. I also managed to figure out how to get the correct date format, which was a lot A LOT of trial and error, but it works well now.